backend/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "risk-dashboard-backend"
version = "0.1.0"
description = "AI-Driven Risk Dashboard Backend"
authors = [{ name = "Risk Dashboard Team" }]
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "sqlalchemy>=2.0.0",
    "alembic>=1.12.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "psycopg2-binary>=2.9.7",
    "redis>=5.0.0",
    "celery>=5.3.0",
    "numpy>=1.24.0",
    "pandas>=2.1.0",
    "scipy>=1.11.0",
    "httpx>=0.25.0",
    "websockets>=12.0",
    "python-jose[cryptography]>=3.3.0",
    "python-multipart>=0.0.6",
    "ccxt>=4.1.0",
    "yfinance>=0.2.0",
    "anthropic>=0.8.0",
    "openai>=1.3.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "httpx>=0.25.0",
    "hypothesis>=6.88.0",
    "mypy>=1.6.0",
    "flake8>=6.1.0",
    "black>=23.9.0",
    "isort>=5.12.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "--cov=app --cov-report=term-missing"

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.black]
line-length = 88
target-version = ['py311']

[tool.isort]
profile = "black"
